Start work on separating plugin API
