+++ plugin-api/src/main/java/net/clgd/ccemux/api/OperatingSystem.java
@@ -1,15 +1,25 @@
-package net.clgd.ccemux;
+package net.clgd.ccemux.api;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Objects;
 
+/**
+ * Used to determine which OS is in use and where application data should be
+ * stored
+ */
 public enum OperatingSystem {
 	Windows(Paths.get(Objects.toString(System.getenv("appdata"), System.getProperty("user.home")))),
 	MacOSX(Paths.get(System.getProperty("user.home")).resolve("Library/Application Support")),
 	Linux(Paths.get(System.getProperty("user.home")).resolve(".local/share")),
 	Other(Paths.get(System.getProperty("user.home")));
 
+	/**
+	 * Gets the directory that should be used for storing application data for this
+	 * OS
+	 * 
+	 * @return The directory that application data should be stored in
+	 */
 	public Path getAppDataDir() {
 		return appDataDir;
 	}
@@ -20,6 +30,11 @@ public enum OperatingSystem {
 		this.appDataDir = appDataDir;
 	}
 
+	/**
+	 * Gets the OS this program is running on
+	 * 
+	 * @return The appropriate value for this OS
+	 */
 	public static OperatingSystem get() {
 		String name = System.getProperty("os.name");
 		if (name.startsWith("Windows"))
