+++ settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'CCEmuX'
+
+include 'plugin-api'
