+++ src/main/java/net/clgd/ccemux/init/Launcher.java
@@ -20,7 +20,7 @@ import dan200.computercraft.ComputerCraft;
 import dan200.computercraft.core.apis.AddressPredicate;
 import lombok.val;
 import lombok.extern.slf4j.Slf4j;
-import net.clgd.ccemux.OperatingSystem;
+import net.clgd.ccemux.api.OperatingSystem;
 import net.clgd.ccemux.emulation.CCEmuX;
 import net.clgd.ccemux.plugins.PluginManager;
 import net.clgd.ccemux.rendering.RendererFactory;
