+++ src/main/java/net/clgd/ccemux/rendering/awt/AWTTerminalFont.java
@@ -1,15 +1,14 @@
 package net.clgd.ccemux.rendering.awt;
 
-import lombok.extern.slf4j.Slf4j;
-import net.clgd.ccemux.Utils;
-import net.clgd.ccemux.rendering.TerminalFont;
-
-import javax.imageio.ImageIO;
-import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
+
+import javax.imageio.ImageIO;
+
+import lombok.extern.slf4j.Slf4j;
+import net.clgd.ccemux.rendering.TerminalFont;
 
 @Slf4j
 public class AWTTerminalFont extends TerminalFont {
