+++ src/main/resources/ccemux.version
@@ -1,2 +1,1 @@
 version=$version
-build=$build
